Hiding "Label All Tasks" button
The default choice is “Label All Tasks” rather than “Label Tasks As Displayed”, which is really confusing when you’re on a tab with your selected tasks (most common labeling setup). With our current setup this code is in the DataManager repo, which we haven’t forked (and is deprecated). Once we update to new version of label studio this will all be in one repo but for now gonna do a hack and just hide the button. Labellers will be able to just click on a task in their tab to get started.

Issue: https://khanacademy.atlassian.net/browse/DI-1500

Test plan:
- ran label studio locally, created a project, imported a task, confirmed button is gone (see screenshot)